cpsc_knowledge_graph/requirements.txt
# CPSC Knowledge Graph Requirements
# ================================

# Core dependencies
pandas>=1.5.0
scikit-learn>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Visualization and dashboard
streamlit>=1.28.0
plotly>=5.14.0
networkx>=3.0

# Optional: For advanced analysis
numpy>=1.0.0
matplotlib>=3.0.0
seaborn>=0.11.0